till 210 - Before new Users table

+++ views.py
@@ -58,8 +58,6 @@ def login():
                 return redirect(url_for('tasks'))
             else:
                 error = 'Invalid username or password.'
-        else:
-            error = 'Both fields are required.'
     return render_template('login.html', form=form, error=error)
 
 
@@ -81,6 +79,7 @@ def logout():
     session.pop('user_id', None)
     flash('Goodbye!')
     return redirect(url_for('login'))
+
 
 @app.route('/tasks/')
 @login_required
@@ -121,25 +120,34 @@ def new_task():
     )
 
 
-
 @app.route('/complete/<int:task_id>/')
 @login_required
 def complete(task_id):
     new_id = task_id
-    db.session.query(Task).filter_by(task_id=new_id).update({"status": "0"})
-    db.session.commit()
-    flash('The task is complete. Nice.')
-    return redirect(url_for('tasks'))
+    task = db.session.query(Task).filter_by(task_id=new_id)
+    if session['user_id'] == task.first().user_id:
+        db.session.query(Task).filter_by(task_id=new_id).update({"status": "0"})
+        db.session.commit()
+        flash('The task is complete. Nice.')
+        return redirect(url_for('tasks'))
+    else:
+        flash('You can only update tasks that belong to you.')
+        return redirect(url_for('tasks'))
 
 
 @app.route('/delete/<int:task_id>/')
 @login_required
 def delete_entry(task_id):
     new_id = task_id
-    db.session.query(Task).filter_by(task_id=new_id).delete()
-    db.session.commit()
-    flash('The task was deleted. Why not add a new one?')
-    return redirect(url_for('tasks'))
+    task = db.session.query(Task).filter_by(task_id=new_id)
+    if session['user_id'] == task.first().user_id:
+        db.session.query(Task).filter_by(task_id=new_id).delete()
+        db.session.commit()
+        flash('The task was deleted. Why not add a new one?')
+        return redirect(url_for('tasks'))
+    else:
+        flash('You can only delete tasks that belong to you.')
+        return redirect(url_for('tasks'))
 
 
 def flash_errors(form):
